- Created MIDI interface blocks - Added strip midi message block - Added MIDI port type - Moved Form blocks to a separated field
