+++ mosaicode_lib_javascript_webaudio/extensions/webaudio/GUI/__init__.py
@@ -1,0 +1,2 @@
+#!/usr/bin/env python
+# __init__
