+++ mosaicode_lib_javascript_webaudio/extensions/webaudio/GUI/number.py
@@ -6,7 +6,7 @@ This module contains the Button class.
 from mosaicode.GUI.fieldtypes import *
 from mosaicode.model.blockmodel import BlockModel
 
-class Text(BlockModel):
+class Number(BlockModel):
 
     # -------------------------------------------------------------------------
     def __init__(self):
@@ -14,30 +14,47 @@ class Text(BlockModel):
 
         self.language = "javascript"
         self.framework = "webaudio"
-        self.help = "Text"
-        self.label = "Text"
+        self.help = "Number"
+        self.label = "Number"
         self.color = "50:150:250:150"
-        self.ports = [{"type":"mosaicode_lib_javascript_webaudio.extensions.ports.string",
+        self.group = "Form"
+
+        self.ports = [{"type":"mosaicode_lib_javascript_webaudio.extensions.ports.float",
                 "label":"Value",
                 "conn_type":"Output",
                 "name":"value"},
-                {"type":"mosaicode_lib_javascript_webaudio.extensions.ports.string",
+                {"type":"mosaicode_lib_javascript_webaudio.extensions.ports.float",
                 "label":"Value",
                 "conn_type":"Input",
                 "name":"invalue"}
             ]
         self.properties = [{"name": "value",
                             "label": "Value",
-                            "type": MOSAICODE_STRING,
-                            "value": ''
+                            "type": MOSAICODE_FLOAT,
+                            "value": '0'
                             },
                            {"name": "label",
                             "label": "Label",
                             "type": MOSAICODE_STRING,
                             "value": "Label"
+                            },
+                            {"name": "min",
+                            "label": "Min",
+                            "type": MOSAICODE_FLOAT,
+                            "lower": 0,
+                             "upper": 20000,
+                            "step": 1,
+                            "value": 0
+                            },
+                            {"name": "max",
+                            "label": "Max",
+                            "type": MOSAICODE_FLOAT,
+                            "lower": 0,
+                             "upper": 20000,
+                            "step": 1,
+                            "value": 100
                             }
                            ]
-        self.group = "GUI"
 
         self.codes["declaration"] = """
 // block_$id$ = $label$
@@ -51,10 +68,7 @@ var $port[invalue]$ = function(value){
 """
 
         self.codes["execution"] = """
-function enter_value$id$(e){
-    e = e || window.event;
-    if (e.keyCode != 13) //Ignore if it is not enter
-        return;
+function change_value$id$(e){
     value = document.getElementById("block_$id$").value;
     for (var i = 0; i < $port[value]$.length ; i++){
         $port[value]$[i](value);
@@ -63,7 +77,6 @@ function enter_value$id$(e){
 """
 
         self.codes["html"] = """
-$prop[label]$ <input type="text" value="$prop[value]$" onkeypress="enter_value$id$(event);"
-id="block_$id$"><br>
+$prop[label]$ <input type="number" min="$prop[min]$" max="$prop[max]$" onchange="change_value$id$(event);" id="block_$id$"><br>
 """
