+++ mosaicode_lib_javascript_webaudio/extensions/webaudio/MIDI/stripmidi.py
@@ -6,7 +6,7 @@ This module contains the Midi2Freq class.
 from mosaicode.GUI.fieldtypes import *
 from mosaicode.model.blockmodel import BlockModel
 
-class Midi2Freq(BlockModel):
+class StripMidi(BlockModel):
 
     # --------------------------------------------------------------------------
     def __init__(self):
@@ -14,30 +14,35 @@ class Midi2Freq(BlockModel):
 
         self.language = "javascript"
         self.framework = "webaudio"
-        self.help = "midi to freq"
-        self.label = "Midi 2 Freq"
-        self.color = "200:200:25:150"
-        self.ports = [{"type":"mosaicode_lib_javascript_webaudio.extensions.ports.float",
+        self.help = "Strip Midi"
+        self.label = "Strip Midi"
+        self.color = "250:250:0:150"
+        self.group = "MIDI"
+
+        self.ports = [{"type":"mosaicode_lib_javascript_webaudio.extensions.ports.midi",
                 "name":"midi_value",
                 "conn_type":"Input",
                 "label":"Midi Value"},
                 {"type":"mosaicode_lib_javascript_webaudio.extensions.ports.float",
-                "label":"Frequency",
+                "label":"Note",
                 "conn_type":"Output",
-                "name":"frequency"}
+                "name":"note"},
+                {"type":"mosaicode_lib_javascript_webaudio.extensions.ports.float",
+                "label":"Velocity",
+                "conn_type":"Output",
+                "name":"velocity"}
             ]
-        self.group = "Conversion"
         self.codes["declaration"] = """
 // block_$id$ = $label$
-var $port[frequency]$ = [];
+var $port[note]$ = [];
+var $port[velocity]$ = [];
 
 var $port[midi_value]$ = function(value){
-    value = (value < 0) ? 0 : value;
-    value = (value >127) ? 127 : value;
-    var arg = ((parseFloat(value) - 69.0) / 12.0);
-    result =  Math.pow(2.0, arg) * 440.0;
-    for (var i = 0; i < $port[frequency]$.length ; i++){
-        $port[frequency]$[i](result);
+    for (var i = 0; i < $port[note]$.length ; i++){
+        $port[note]$[i](value[0]);
+    }
+    for (var i = 0; i < $port[note]$.length ; i++){
+        $port[velocity]$[i](value[1]);
     }
     return true;
     };
